src/cogs/requirements.txt
discord.py==1.6.0
validators==0.18.2
requests==2.25.1
discord==1.7.3
Pillow==8.3.2
